Refactor pipeline artifacts into module
