+++ src/framework.rs
@@ -11,33 +11,6 @@ use serde_json::Value as JsonValue;
 pub type Error = Box<dyn std::error::Error>;
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct ChainWellKnownInfo {
-    pub shelley_known_slot: u64,
-    pub shelley_known_hash: String,
-    pub shelley_known_time: u64,
-}
-
-impl ChainWellKnownInfo {
-    pub fn try_from_magic(magic: u64) -> Result<ChainWellKnownInfo, Error> {
-        match magic {
-            MAINNET_MAGIC => Ok(ChainWellKnownInfo {
-                shelley_known_slot: 4492799,
-                shelley_known_hash:
-                    "f8084c61b6a238acec985b59310b6ecec49c0ab8352249afd7268da5cff2a457".to_string(),
-                shelley_known_time: 1596059071,
-            }),
-            TESTNET_MAGIC => Ok(ChainWellKnownInfo {
-                shelley_known_slot: 1598399,
-                shelley_known_hash:
-                    "7e16781b40ebf8b6da18f7b5e8ade855d6738095ef2f1c58c77e88b6e45997a4".to_string(),
-                shelley_known_time: 1595967596,
-            }),
-            _ => Err("can't infer well-known chain infro from specified magic".into()),
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
 #[serde(rename_all = "snake_case")]
 pub enum MetadatumRendition {
     MapJson(JsonValue),
@@ -258,56 +231,4 @@ pub struct Event {
     pub data: EventData,
 
     pub fingerprint: Option<String>,
-}
-
-pub type PartialBootstrapResult = Result<(JoinHandle<()>, StageReceiver), Error>;
-
-pub type BootstrapResult = Result<JoinHandle<()>, Error>;
-
-pub trait SourceConfig {
-    fn bootstrap(&self) -> PartialBootstrapResult;
-}
-
-pub trait FilterConfig {
-    fn bootstrap(&self, input: StageReceiver) -> PartialBootstrapResult;
-}
-
-pub trait SinkConfig {
-    fn bootstrap(&self, input: StageReceiver) -> BootstrapResult;
-}
-
-pub type StageReceiver = std::sync::mpsc::Receiver<Event>;
-
-pub type StageSender = std::sync::mpsc::SyncSender<Event>;
-
-/// The amount of events an inter-stage channel can buffer before blocking
-///
-/// If a filter or sink has a consumption rate lower than the rate of event
-/// generations from a source, the pending events will buffer in a queue
-/// provided by the corresponding mpsc channel implementation. This constant
-/// defines the max amount of events that the buffer queue can hold. Once
-/// reached, the previous stages in the pipeline will start blockin on 'send'.
-///
-/// This value has a direct effect on the amount of memory consumed by the
-/// process. The higher the buffer, the higher potential memory consumption.
-///
-/// This value has a direct effect on performance. To allow _pipelining_
-/// benefits, stages should be allowed certain degree of flexibility to deal
-/// with resource constrains (such as network or cpu). The lower the buffer, the
-/// lower degree of flexibility.
-const DEFAULT_INTER_STAGE_BUFFER_SIZE: usize = 1000;
-
-pub type StageChannel = (StageSender, StageReceiver);
-
-/// Centralizes the implementation details of inter-stage channel creation
-///
-/// Concrete channel implementation is subject to change. We're still exploring
-/// sync vs unbounded and threaded vs event-loop. Until we have a long-term
-/// strategy, it makes sense to have a single place in the codebase that can be
-/// used to change from one implementation to the other without incurring on
-/// heavy refactoring throughout several files.
-///
-/// Sometimes centralization is not such a bad thing :)
-pub fn new_inter_stage_channel(buffer_size: Option<usize>) -> StageChannel {
-    std::sync::mpsc::sync_channel(buffer_size.unwrap_or(DEFAULT_INTER_STAGE_BUFFER_SIZE))
 }
