+++ src/lib.rs
@@ -1,6 +1,11 @@
+mod framework;
+mod utils;
+pub mod config;
+pub mod pipelining;
 pub mod filters;
-pub mod framework;
 pub mod mapper;
 pub mod sinks;
 pub mod sources;
-pub mod utils;
+
+
+pub type Error = Box<dyn std::error::Error>;
