+++ src/sources/common.rs
@@ -11,7 +11,7 @@ use pallas::ouroboros::network::{
 use serde::{de::Visitor, Deserializer};
 use serde_derive::{Deserialize, Serialize};
 
-use crate::framework::{ChainWellKnownInfo, Error};
+use crate::{config::ChainWellKnownInfo, Error};
 
 #[derive(Debug, Deserialize)]
 pub enum BearerKind {
