+++ src/sources/mod.rs
@@ -1,3 +1,6 @@
-pub mod common;
+mod common;
+
 pub mod n2c;
 pub mod n2n;
+
+pub use common::*;
